#29-refactor: Planner 관련 도메인들 모두 planner로 이름 통일
