+++ src/main/java/com/example/demo/service/planner/TripDateNotFoundException.java
@@ -1,4 +1,4 @@
-package com.example.demo.service.tripPlanner;
+package com.example.demo.service.planner;
 
 public class TripDateNotFoundException extends RuntimeException {
     public TripDateNotFoundException(String message) {
