+++ src/main/java/com/example/demo/service/planner/TripNotFoundException.java
@@ -1,4 +1,4 @@
-package com.example.demo.service.tripPlanner;
+package com.example.demo.service.planner;
 
 public class TripNotFoundException extends RuntimeException {
     public TripNotFoundException(String message) {
